Reported by jorrit@ncode.nl, Feb 05, 2010
The file modification time is read from the file and saved in the 
FileHandlerCacheItem. It is then compared to the supplied If-Modified-
Since header.
The last modification date is fetched using 
file.LastWriteTime.ToUniversalTime(). However, the returned time has a sub-
second precision.
For instance, one file I tested was last modified at 13-1-2010 10:30:58. 
The DateTime object associated with the cache entry has a ToString() value 
of "13-1-2010 10:30:58" and the Ticks property has the value 
633989754586086250. When a browser sets the If-Modified-Since field to 13-
1-2010 10:30:58, another DateTime instance is created, but this one has a 
Ticks value of 633989754580000000. While you would expect that both 
DateTime instances are equal, they are in fact not equal as the last 
modification time is slightly larger than the if-modified-since value.

I think the library should cut off or round the sub-seconds from the 
DateTime object while reading the last modification time of a file.